Task #53: Move Meetup logic to Event
