+++ spec/Event/Entity/EventEntitySpec.php
@@ -5,6 +5,7 @@ namespace spec\Event\Entity;
 use Event\Entity\EventEntity;
 use Geo\Entity\LocationEntity;
 use PhpSpec\ObjectBehavior;
+use User\Entity\UserEntity;
 
 class EventEntitySpec extends ObjectBehavior
 {
@@ -29,5 +30,26 @@ class EventEntitySpec extends ObjectBehavior
         $this->getTitle()->shouldReturn('Event title');
         $this->getDescription()->shouldReturn('Event description');
         $this->getLocation()->shouldReturn($location);
+    }
+
+    public function it_will_add_attendee_to_event(UserEntity $attendee)
+    {
+        $this->addAttendee($attendee);
+
+        $this->getAttendees()->shouldReturn([$attendee]);
+    }
+
+    public function it_will_add_user_to_not_coming_list(UserEntity $notComing)
+    {
+        $this->addNotComing($notComing);
+
+        $this->getNotComingList()->shouldReturn([$notComing]);
+    }
+
+    public function it_will_add_user_to_maybe_coming_list(UserEntity $maybeComing)
+    {
+        $this->addMaybeComing($maybeComing);
+
+        $this->getMaybeComingList()->shouldReturn([$maybeComing]);
     }
 }
